chore(NO-US): Add prettier to enforce style consistency
Prettier allow us to write code in our personal style and it will always format the code
consistently. This removes the burden of the developer to conform to a predefined styleguide, and
reduces the amount of conflicts due to stylistic(and most often insignificant) changes like
whitespace, commans and semicolons.